requirements.txt
pytest
python-socketio
python-dotenv
aiohttp
